refactor(desktop): share protocol and split native shell
